pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "blockchain-agent-md"
version = "0.1.0"
description = "Autonomous agents that securely write, audit, and deploy smart contracts"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Avasis AI", email = "contact@avasis.ai" }
]
maintainers = [
    { name = "Avasis AI", email = "contact@avasis.ai" }
]
keywords = [
    "blockchain",
    "smart-contracts",
    "solidity",
    "web3",
    "audit",
    "security",
    "agent",
    "ethereum",
    "cryptocurrency"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Security",
    "Topic :: Utilities",
]
requires-python = ">=3.8"
dependencies = [
    "web3>=6.0",
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.urls]
Homepage = "https://github.com/avasis-ai/blockchain-agent-md"
Documentation = "https://github.com/avasis-ai/blockchain-agent-md#readme"
Repository = "https://github.com/avasis-ai/blockchain-agent-md.git"
Issues = "https://github.com/avasis-ai/blockchain-agent-md/issues"

[project.scripts]
blockchain-agent = "blockchain_agent_md.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/blockchain_agent_md"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.isort]
profile = "black"
multi_line_output = 3
